Add markup  and styles for page structure
