+++ src/index.html
@@ -7,6 +7,17 @@
     <title>Document</title>
 </head>
 <body>
-    <ul class="to-do-list"></ul>
+    <div class="container">
+        <div class="today">
+            <p>Today's To Do</p>
+            <i class="fa-solid fa-arrows-rotate"></i>
+        </div>
+        <form>
+            <input type="text" placeholder="Add to your list..." class="list-input" id="list-add">
+          <ul class="to-do-list"></ul>
+          <button type="button">Clear all completed</button>
+        </form>
+    </div>
+ 
 </body>
 </html>
